fix(browser): action.result discriminator + result-parse resilience + WS send serialization + snapshot uuid
Autonomous Playwright harness (.browser-e2e) drove the real extension and
surfaced a chain of cross-runtime/runtime bugs after navigate started working:

- ActionExecutor: stamp `kind` onto every success payload — server's
  ActionSuccessPayload uses Jackson NAME dispatch on `kind`; without it the
  server threw InvalidTypeIdException, closed the WS (1011) and the action
  completed as SESSION_DETACHED.
- EdgeWebSocketHandler.onActionResult: catch deserialization failures and
  deliver a typed RESULT_PARSE_ERROR instead of letting the exception close the
  socket (one bad result must not detach the whole session).
- BrowserSessionRegistry: wrap each session in ConcurrentWebSocketSessionDecorator
  + route handler replies through session.getWs() — Tomcat WS forbids concurrent
  sends; heartbeat-ack racing action.execute caused TEXT_PARTIAL_WRITING/SEND_FAILED.
- SnapshotRequestHandler.uuid: `(deps.uuid ?? crypto.randomUUID)()` called
  randomUUID detached → "Illegal invocation" on every observe → snapshot timeout.
  Call it with its crypto receiver.

navigate now drives the real browser end-to-end (ok:true) verified by the harness.

+++ mateclaw-server/src/main/java/vip/mate/browser/orchestrator/ActionExecutionService.java
@@ -263,7 +263,11 @@ public class ActionExecutionService {
                         "tab_ref", mapper.convertValue(req.tabRef(), Object.class),
                         "kind", req.kind().wire(),
                         "params", mapper.convertValue(req.params(), MAP_TYPE),
-                        "deadline_ms", req.deadlineMs()))
+                        // int, not long: the global ObjectMapper stringifies Long/long
+                        // (Snowflake-ID guard in JacksonConfig), but the extension's
+                        // parseActionRequest requires deadline_ms to be a JSON *number*.
+                        // Deadlines are seconds-to-minutes, far within int range.
+                        "deadline_ms", Math.toIntExact(req.deadlineMs())))
                 .build();
         session.getWs().sendMessage(new TextMessage(mapper.writeValueAsString(message)));
     }
